# Conflicts:
#	src/main/java/fr/ensicaen/genielogiciel/mvp/presenter/GamePresenter.java

+++ src/main/java/fr/ensicaen/genielogiciel/mvp/model/map/wind/Wind.java
@@ -1,5 +1,5 @@
 package fr.ensicaen.genielogiciel.mvp.model.map.wind;
 public interface Wind {
-    int getWindKnot();
+    double getWindKnot();
     WindDirection getWindDirection();
 }
